Move initializers loading to container
