+++ lib/hanami/loader.rb
@@ -31,7 +31,6 @@ module Hanami
         load_configuration_load_paths!
         load_rack!
         load_frameworks!
-        load_initializers!
       end
     end
 
@@ -247,10 +246,5 @@ module Hanami
       configuration.namespace || application_module
     end
 
-    def load_initializers!
-      Dir["#{configuration.root}/config/initializers/**/*.rb"].each do |file|
-        require file
-      end
-    end
   end
 end
