+++ test/fixtures/tea_shop/config.ru
@@ -1,3 +1,3 @@
 require './config/environment'
 
-run Hanami::Container.new
+run TeaShop::Container.new
